requirements.txt
numpy~=1.24.2
mediapipe~=0.9.3.0
opencv-python~=4.7.0
pyautogui~=0.9.48
pytest~=7.3.1
pyyaml~=6.0